Implement Secure OAuth2 Authentication for Enhanced API Access
In response to the recent API URL change, it has become mandatory to incorporate OAuth2 authentication.
To align with Vulndb's recommendations and ensure a secure connection, we have implemented the retrieval
of an access token using our client credentials and secret.

Considering that HTTPS handles URL signing automatically, there is no need for explicit URL signing in this implementation.
The security aspects are effectively managed by the HTTPS protocol.

Additionally, we have retained the inclusion of a User Agent in our GET request to provide comprehensive information about our VulnDB Data Mirror application.

Signed-off-by: Andres Tito <andres.tito@rohde-schwarz.com>